% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aov_pipe_pval}
\alias{aov_pipe_pval}
\title{Anova p value pipe
Allows the extraction of an ANOVA p value in a pipe}
\usage{
aov_pipe_pval(data, variable, group)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{variable}{A character string corresponding to the studied variable.}

\item{group}{A character string corresponding to the comparative groups.}
}
\description{
Anova p value pipe
Allows the extraction of an ANOVA p value in a pipe
}
