% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{autoplot}
\alias{autoplot}
\title{Automatised plot
This is a wrapper function to set a plot with different parameters from ggplot.}
\usage{
autoplot(
  data,
  variable,
  group,
  group_filter_vector = NULL,
  na_exclude_group = T,
  plot_type = "Boxplot",
  add_points = T,
  error_bar = "IC95",
  stat = "param"
)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{variable}{A character string corresponding to the studied variable.}

\item{group}{A character string corresponding to the comparative groups.}

\item{group_filter_vector}{A character string/vector defining groups sublevels to be included in the analysis.}

\item{na_exclude_group}{Should missing values be excluded from groups?}

\item{plot_type}{Set the plot type (Boxplot, Barchart_mean or  Barchart_count).}

\item{add_points}{Add data points as a scatter plot to the graph.}

\item{error_bar}{Type of error bar to be shown on the barcharts (IC95 or hide).}

\item{stat}{Type of statistics to show (param, non_param or no).}
}
\description{
Automatised plot
This is a wrapper function to set a plot with different parameters from ggplot.
}
