% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{autoplot_paired}
\alias{autoplot_paired}
\title{Automatised plot for paired data
This is a wrapper function to set a plot with different parameters from ggplot for paired data.}
\usage{
autoplot_paired(
  data,
  timepoints,
  group,
  plot_type = "Mean_lines",
  add_points = F,
  add_lines = T,
  add_individual_lines = F,
  error_bar = "IC95",
  stat = "param",
  alpha_line = 0.5
)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{timepoints}{A character vector corresponding to the studied paired variables.}

\item{group}{A character string corresponding to the comparative groups.}

\item{plot_type}{Set the plot type (Boxplot, Barchart_mean or  Barchart_count).}

\item{add_points}{Add data points as a scatter plot to the graph.}

\item{add_lines}{Add line to paired data}

\item{add_individual_lines}{Add each individual (samples) lines to the plot.}

\item{error_bar}{Type of error bar to be shown on the barcharts (IC95 or se or hide).}

\item{stat}{Type of statistics to show (param, non_param or no).}

\item{alpha_line}{Transparency of the paired lines}
}
\description{
Automatised plot for paired data
This is a wrapper function to set a plot with different parameters from ggplot for paired data.
}
