% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{data_aml}
\alias{data_aml}
\title{Global parameters}
\format{
An object of class \code{data.frame} with 151 rows and 21 columns.
}
\usage{
data_aml
}
\description{
Global parameters
}
\keyword{datasets}
