% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{descriptive_table}
\alias{descriptive_table}
\title{All-in-one publication ready descriptive table
This function output a publication-ready descriptive table from a dataset. Each columns of the dataset is compared between the defined groups.}
\usage{
descriptive_table(
  data,
  group,
  na.include = F,
  percent_type = 1,
  padj_method = "none",
  show_methods = F,
  exclude_vector = c("Patient_id", "patient_id", "Sample_ID", "Whole_cohort")
)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{group}{A character string corresponding to the comparative groups.}

\item{na.include}{Should missing values be included in groups?}

\item{percent_type}{For categorical variables, should percentage be calculated on the row (1) or columns (2)?}

\item{padj_method}{Select the p.value adjustment method (none, fdr, holm or bonferonni).}

\item{show_methods}{Should statistical tests names be precised in a supplementary column?}

\item{exclude_vector}{columns to exclude from the analysis.}
}
\description{
All-in-one publication ready descriptive table
This function output a publication-ready descriptive table from a dataset. Each columns of the dataset is compared between the defined groups.
}
