% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{regression_dataframes}
\alias{regression_dataframes}
\title{Regression base and diagnosis dataframe
Provides augmented data for regression diagnosis}
\usage{
regression_dataframes(y_var, x_var, data, model = "lm", cor_type = "pearson")
}
\arguments{
\item{y_var}{A character string corresponding to the y variable.}

\item{x_var}{A character string corresponding to the x variable(s).}

\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{model}{Type of generalised linear model (lm or gam).}

\item{cor_type}{Type of correlation to show on the graph (pearson or spearman).}
}
\description{
Regression base and diagnosis dataframe
Provides augmented data for regression diagnosis
}
