% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{regression_table_multi_cox}
\alias{regression_table_multi_cox}
\title{All-in-one publication ready regression table for cox models
This function output a publication-ready regression table from a dataset. Each column of the dataset is included as a x variable in a multivariate model.}
\usage{
regression_table_multi_cox(data, y_var, time_var)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{y_var}{A character string corresponding to the y variable.}

\item{family}{Generalized linear model family (gaussian or binomial)}
}
\description{
All-in-one publication ready regression table for cox models
This function output a publication-ready regression table from a dataset. Each column of the dataset is included as a x variable in a multivariate model.
}
