% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ROC_calc}
\alias{ROC_calc}
\title{Dedicated function to plot ROC Curves & compute parameters
This function output individual ROC Curves for each x-variable provided.}
\usage{
ROC_calc(data, y_var, x_var, value = "raw")
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{y_var}{A character string corresponding to the y variable. !!Must be a 2-level factor (TRUE/FALSE, 1/0)}

\item{x_var}{A vector of numeric variable to test}

\item{value}{Use raw parameters value or predictions from logit model?}
}
\description{
Dedicated function to plot ROC Curves & compute parameters
This function output individual ROC Curves for each x-variable provided.
}
