% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ROC_curves}
\alias{ROC_curves}
\title{Dedicated function to plot ROC Curves
This function output individual ROC Curves for each x-variable provided.}
\usage{
ROC_curves(data, y_var, x_var)
}
\arguments{
\item{data}{A dataframe with row corresponding to samples/patients and columns to variables.}

\item{y_var}{A character string corresponding to the y variable. !!Must be a 2-level factor (TRUE/FALSE, 1/0)}

\item{time_var}{A vector of numeric variable to test}
}
\description{
Dedicated function to plot ROC Curves
This function output individual ROC Curves for each x-variable provided.
}
